// ClientApp/app/models/vehicle.ts
export interface KeyPairValue {
    id: number;
    name: string;
}

export interface Contact {
    name: string;
    phone: string;
    email: string;
}

export interface Vehicle {
    id: number;
    model: KeyPairValue;
    make: KeyPairValue;
    isRegistered: boolean;
    features: KeyPairValue[];
    contact: Contact;
    lastUpdate: string;
}

export interface SaveVehicle {
    id: number;
    modelId: number;
    makeId: number;
    isRegistered: boolean;
    features: number[];
    contact: Contact;
}